+++ Source/SwiftLintFramework/Rules/MultilineArgumentsConfiguration.swift
@@ -1,7 +1,9 @@
 import Foundation
 
 private enum ConfigurationKey: String {
+    case severity = "severity"
     case firstArgumentLocation = "first_argument_location"
+    case onlyEnforceAfterFirstClosureOnFirstLine = "only_enforce_after_first_closure_on_first_line"
 }
 
 public struct MultilineArgumentsConfiguration: RuleConfiguration, Equatable {
@@ -23,23 +25,38 @@ public struct MultilineArgumentsConfiguration: RuleConfiguration, Equatable {
 
     private(set) var severityConfiguration = SeverityConfiguration(.warning)
     private(set) var firstArgumentLocation = FirstArgumentLocation.anyLine
+    private(set) var onlyEnforceAfterFirstClosureOnFirstLine = false
 
     public var consoleDescription: String {
+        // swiftlint:disable line_length
         return severityConfiguration.consoleDescription +
-            ", \(ConfigurationKey.firstArgumentLocation.rawValue): \(firstArgumentLocation.rawValue)"
+            ", \(ConfigurationKey.firstArgumentLocation.rawValue): \(firstArgumentLocation.rawValue)" +
+            ", \(ConfigurationKey.onlyEnforceAfterFirstClosureOnFirstLine.rawValue): \(onlyEnforceAfterFirstClosureOnFirstLine)"
+        // swiftlint:enable line_length
     }
 
     public mutating func apply(configuration: Any) throws {
+        let error = ConfigurationError.unknownConfiguration
+
         guard let configuration = configuration as? [String: Any] else {
-            throw ConfigurationError.unknownConfiguration
+            throw error
         }
 
-        if let modeString = configuration[ConfigurationKey.firstArgumentLocation.rawValue] {
-            try firstArgumentLocation = FirstArgumentLocation(value: modeString)
-        }
+        for (string, value) in configuration {
+            guard let key = ConfigurationKey(rawValue: string) else {
+                throw error
+            }
 
-        if let severityString = configuration["severity"] as? String {
-            try severityConfiguration.apply(configuration: severityString)
+            switch (key, value) {
+            case (.firstArgumentLocation, _):
+                try firstArgumentLocation = FirstArgumentLocation(value: value)
+            case (.severity, let stringValue as String):
+                try severityConfiguration.apply(configuration: stringValue)
+            case (.onlyEnforceAfterFirstClosureOnFirstLine, let boolValue as Bool):
+                onlyEnforceAfterFirstClosureOnFirstLine = boolValue
+            default:
+                throw error
+            }
         }
     }