+++ CHANGELOG.md
@@ -17,6 +17,11 @@
   changed.  
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#2249](https://github.com/realm/SwiftLint/issues/2249)
+  
+  * Add `only_enforce_after_first_closure_on_first_line` configuration
+  to `multiline_arguments`  
+  [Mike Ciesielka](https://github.com/maciesielka)
+  [#1896](https://github.com/realm/SwiftLint/issues/1896)
 
 * Add `anyobject_protocol` opt-in rule which suggests using `AnyObject` over
   `class` for class-only protocols.